Add shouldCleanup option

+++ ordering-keys-prober/src/main/java/com/google/cloud/pubsub/prober/Prober.java
@@ -80,6 +80,7 @@ public class Prober {
     String endpoint = new String("pubsub.googleapis.com:443");
     String topicName = new String("");
     String subscriptionName = new String("");
+    boolean shouldCleanup = true;
     SubscriptionType subscriptionType = SubscriptionType.STREAMING_PULL;
     double messageFailureProbability = 0.0;
     long publishFrequency = 1_000_000L;
@@ -102,6 +103,11 @@ public class Prober {
 
     public Builder setEndpoint(String endpoint) {
       this.endpoint = endpoint;
+      return this;
+    }
+
+    public Builder setShouldCleanup(boolean shouldCleanup) {
+      this.shouldCleanup = shouldCleanup;
       return this;
     }
 
@@ -198,6 +204,7 @@ public class Prober {
 
   private final String project;
   private final String endpoint;
+  private final boolean shouldCleanup;
   private final String topicName;
   private final String subscriptionName;
   private final SubscriptionType subscriptionType;
@@ -256,6 +263,7 @@ public class Prober {
     this.subscriptionType = builder.subscriptionType;
     this.subscriptionName = builder.subscriptionName;
     this.topicName = builder.topicName;
+    this.shouldCleanup = builder.shouldCleanup;
     this.endpoint = builder.endpoint;
     this.project = builder.project;
     subscribers = new Subscriber[subscriberCount];
@@ -655,6 +663,7 @@ public class Prober {
 
   // Returns true if a topic or subscription was deleted.
   private boolean cleanup() {
+    if (!shouldCleanup) return false;
     boolean deleted = deleteSubscription(fullSubscriptionName);
     deleted = deleted || deleteTopic(fullTopicName);
     return deleted;
